+++ src/main/java/com/tl/myspring/model/BookRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 /**
  * Basic repository for CRUD operations on {@see Book books}.
  */
-public interface BookRepository extends PagingAndSortingRepository<Book, Long>{
+public interface BookRepository extends PagingAndSortingRepository<Book, Integer>{
 
 //  /**
 //   * Find a <code>Book</code> by ID.
